Refactored awesome books to use modules
