+++ index.html
@@ -5,7 +5,7 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
 
-    <link rel="stylesheet" href="./styles/style.css" />
+    <link rel="stylesheet" href="index.css" />
     <title>Awesome Books</title>
   </head>
   <body>
@@ -55,9 +55,8 @@
          </ul>
       </div>
       <div class="book-wrapper">
-        <h2>My<span class = "style-bl">Book</span>List</h2>
-
         <table class = "book-table">
+          <h2>My<span class = "style-bl">Book</span>List</h2>
           <tbody id = "table-body">
           </tbody>
            
@@ -83,6 +82,6 @@
     </footer>
 
 
-    <script src="awesome.js"></script>
+    <script type = "module" src="index.js"></script>
   </body>
 </html>
