+++ modules/createBooks.js
@@ -1,0 +1,6 @@
+export default class BookStore {
+    constructor(title, author) {
+      this.title = title;
+      this.author = author;
+    }
+  }
